+++ it-play-30/src/test/scala/uk/gov/hmrc/govukfrontend/support/TemplateIntegrationSpec.scala
@@ -32,9 +32,8 @@ import scala.reflect.ClassTag
   */
 abstract class TemplateIntegrationSpec[T: OWrites: Arbitrary, C <: Template1[T, HtmlFormat.Appendable]: ClassTag](
   govukComponentName: String,
-  seed: Option[String] = None,
-  maximumCompression: Boolean = false
-) extends TemplateIntegrationBaseSpec[T](govukComponentName, seed, maximumCompression) {
+  seed: Option[String] = None
+) extends TemplateIntegrationBaseSpec[T](govukComponentName, seed) {
 
   protected val libraryName: String = "govuk"